pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bist-swing-system"
version = "0.1.0"
description = "BIST swing trading system (backtest + live alerts)"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "yfinance>=0.2.37",
  "matplotlib>=3.7",
  "pyyaml>=6.0",
  "requests>=2.31",
  "python-dotenv>=1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.pyright]
typeCheckingMode = "basic"
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
